requirements.txt
aiohttp==3.7.3
gevent_socketio==0.3.6
